+++ package.json
@@ -31,8 +31,8 @@
     "distdocs": "cd docs && node node_modules/gatsby/cli.js build --prefix-paths --no-uglify",
     "apidocs": "node node_modules/@microsoft/api-extractor/bin/api-extractor run --local && node node_modules/@microsoft/api-documenter/bin/api-documenter markdown --input-folder etc --output-folder built/apidocs",
     "watch": "node node_modules/rollup/dist/bin/rollup -c rollup.config.ts -w",
-    "test": "node node_modules/ts-mocha/bin/ts-mocha -p ./tsconfig.spec.json src/**/*.spec.ts",
-    "testwatch": "node node_modules/ts-mocha/bin/ts-mocha -p ./tsconfig.spec.json src/**/*.spec.ts -w --watch-extensions ts",
+    "test": "node node_modules/ts-mocha/bin/ts-mocha -p ./tests/tsconfig.json tests/**/*.spec.ts",
+    "testwatch": "node node_modules/ts-mocha/bin/ts-mocha -p ./tests/tsconfig.json tests/**/*.spec.ts -w --watch-extensions ts",
     "tools": "npx http-server . -c-1",
     "installdocs": "cd docs && npm install --force",
     "installdev": "npm install && npm run installdocs",
